refactor(dashboard): split 3 largest components (731/731/712 -> 10 files)
governance.ts (731) -> 4 files:
  - governance-store.ts (147): signals + async actions
  - governance-utils.ts (146): pure label/format functions
  - governance-panels.ts (310): GuardrailPane, RuntimeParams, ActivityRail
  - governance.ts (207): main layout + toolbar + inbox

command/swarm.ts (731) -> 3 files:
  - swarm-cards.ts (156): worker/blocker/checklist cards
  - swarm-storyboard.ts (302): lane strips, health bar, timeline
  - swarm.ts (293): SwarmPanel + SwarmSurface

mission-cards.ts (712) -> 3 files:
  - mission-session-cards.ts (241): session brief + detail
  - mission-agent-cards.ts (193): agent/keeper/signal cards
  - mission-cards.ts (310): context bar, briefing, attention

All re-exports preserved for existing consumers. Vite build passes.

Ref #1518
