+++ dashboard/src/components/command/swarm.ts
@@ -1,35 +1,15 @@
 import { html } from 'htm/preact'
-import type {
-  CommandPlaneRunResolutionRecommendation,
-  CommandPlaneRunResolutionState,
-  CommandPlaneSwarmBlocker,
-  CommandPlaneSwarmChecklistItem,
-  CommandPlaneSwarmFlag,
-  CommandPlaneSwarmGap,
-  CommandPlaneSwarmLane,
-  CommandPlaneSwarmProof,
-  CommandPlaneSwarmResponse,
-  CommandPlaneSwarmTimelineEvent,
-  CommandPlaneSwarmWorker,
-} from '../../types'
 import {
   commandPlaneSwarm,
   commandPlaneSwarmError,
   commandPlaneSwarmLoading,
 } from '../../command-store'
-import {
-  confirmOperatorPendingAction,
-  dispatchOperatorAction,
-  operatorActionBusy,
-  operatorSnapshot,
-} from '../../operator-store'
 import { route } from '../../router'
 import { PanelSemanticDetails } from '../common/semantic-layer'
 import { workflowContextForRoute } from '../../workflow-context'
 import { ProvenanceChip } from '../common/provenance-strip'
 import {
   currentCommandPlaneSummary,
-  dashboardActorName,
   dashboardSwarmOperationId,
   dashboardSwarmRunId,
   relativeTime,
@@ -38,369 +18,17 @@ import {
 } from './helpers'
 import { formatMessageContent } from '../ops/helpers'
 import { TraceRow } from './topology'
+import { SwarmBlockerCard, SwarmChecklistCard, SwarmEventNode, SwarmGapDot, SwarmProofPanel, SwarmWorkerCard } from './swarm-cards'
+import {
+  SwarmHealthBar,
+  SwarmLaneStrip,
+  SwarmRunResolutionCard,
+  SwarmStoryboard,
+} from './swarm-storyboard'
 
-function previewText(value: unknown): string {
-  if (typeof value === 'string') return value
-  if (value == null) return ''
-  try {
-    return JSON.stringify(value, null, 2)
-  } catch {
-    return String(value)
-  }
-}
-
-function runResolutionTone(
-  recommendation: CommandPlaneRunResolutionRecommendation | null | undefined,
-  resolution: CommandPlaneRunResolutionState | null | undefined,
-): string {
-  if (resolution?.status === 'abandoned') return 'warn'
-  if (recommendation?.recommended_kind === 'continue') return 'warn'
-  if (recommendation?.recommended_kind === 'rerun') return 'bad'
-  return 'ok'
-}
-
-function runResolutionLabel(kind?: string | null): string {
-  switch (kind) {
-    case 'continue':
-    case 'continued':
-      return '계속'
-    case 'rerun':
-      return '재실행'
-    case 'abandon':
-    case 'abandoned':
-      return '포기'
-    default:
-      return kind?.trim() || '결정'
-  }
-}
-
-export function SwarmRunResolutionCard({ swarm }: { swarm: CommandPlaneSwarmResponse }) {
-  const runId = swarm.run_id
-  const recommendation = swarm.resolution_recommendation
-  const resolution = swarm.run_resolution
-  if (!runId || (!recommendation && !resolution)) return null
-
-  const actor = dashboardActorName() ?? 'dashboard'
-  const pendingConfirm =
-    operatorSnapshot.value?.pending_confirms.find(item =>
-      item.target_type === 'swarm_run' && item.target_id === runId,
-    ) ?? null
-  const tone = runResolutionTone(recommendation, resolution)
-  const operationId = swarm.operation?.operation_id ?? swarm.operation_id ?? undefined
-  const basePayload: Record<string, unknown> = {
-    run_id: runId,
-  }
-  if (operationId) basePayload.operation_id = operationId
-  if (recommendation?.reason) basePayload.reason = recommendation.reason
-
-  const previewAction = async (actionType: 'swarm_run_continue' | 'swarm_run_rerun' | 'swarm_run_abandon') => {
-    await dispatchOperatorAction({
-      actor,
-      action_type: actionType,
-      target_type: 'swarm_run',
-      target_id: runId,
-      payload: basePayload,
-    })
-  }
-
-  const confirmPending = async (decision: 'confirm' | 'deny') => {
-    if (!pendingConfirm) return
-    await confirmOperatorPendingAction(actor, pendingConfirm.confirm_token, decision)
-  }
-
-  return html`
-    <article class="command-guide-card ${toneClass(tone)}">
-      <div class="command-guide-head">
-        <strong>Run Resolution</strong>
-        <span class="command-chip ${toneClass(tone)}">
-          ${runResolutionLabel(resolution?.status ?? recommendation?.recommended_kind ?? null)}
-        </span>
-      </div>
-      <p>
-        ${resolution?.status === 'abandoned'
-          ? `이 run은 ${resolution.decided_by}가 ${relativeTime(resolution.decided_at)}에 soft abandon 처리했습니다. ${resolution.reason}`
-          : recommendation?.reason ?? '이 run에 대한 별도 resolution recommendation은 아직 없습니다.'}
-      </p>
-      <div class="command-card-grid">
-        <span>Run</span><span>${runId}</span>
-        <span>Provenance</span><span><${ProvenanceChip} item=${{ kind: recommendation?.provenance ?? 'recorded' }} /></span>
-        <span>Engine</span><span>${recommendation?.decision_engine ?? 'operator_record'}</span>
-        <span>Authoritative</span><span>${recommendation?.authoritative ? 'yes' : 'no'}</span>
-      </div>
-      ${recommendation?.evidence
-        ? html`
-            <div class="command-tag-row">
-              <span class="command-tag">joined ${recommendation.evidence.joined_workers ?? 0}</span>
-              <span class="command-tag">trace ${recommendation.evidence.trace_events ?? 0}</span>
-              <span class="command-tag">message ${recommendation.evidence.message_events ?? 0}</span>
-              ${recommendation.evidence.runtime_blocker
-                ? html`<span class="command-tag ${toneClass('bad')}">${recommendation.evidence.runtime_blocker}</span>`
-                : null}
-            </div>
-          `
-        : null}
-      ${pendingConfirm
-        ? html`
-            <div class="command-guide-card warn">
-              <div class="command-guide-head">
-                <strong>확인 대기</strong>
-                <span class="command-chip warn">${pendingConfirm.confirm_token}</span>
-              </div>
-              ${pendingConfirm.preview ? html`<pre class="command-trace-detail">${previewText(pendingConfirm.preview)}</pre>` : null}
-              <div class="command-action-row">
-                <button class="control-btn" onClick=${() => { void confirmPending('confirm') }} disabled=${operatorActionBusy.value}>확인 실행</button>
-                <button class="control-btn ghost" onClick=${() => { void confirmPending('deny') }} disabled=${operatorActionBusy.value}>취소</button>
-              </div>
-            </div>
-          `
-        : recommendation
-          ? html`
-              <div class="command-action-row">
-                ${recommendation.continue_available
-                  ? html`<button class="control-btn ghost" onClick=${() => { void previewAction('swarm_run_continue') }} disabled=${operatorActionBusy.value}>Continue</button>`
-                  : null}
-                ${recommendation.rerun_available
-                  ? html`<button class="control-btn" onClick=${() => { void previewAction('swarm_run_rerun') }} disabled=${operatorActionBusy.value}>Rerun</button>`
-                  : null}
-                ${recommendation.abandon_available
-                  ? html`<button class="control-btn ghost" onClick=${() => { void previewAction('swarm_run_abandon') }} disabled=${operatorActionBusy.value}>Abandon</button>`
-                  : null}
-              </div>
-            `
-          : null}
-    </article>
-  `
-}
-
-function swarmLaneTone(lane: CommandPlaneSwarmLane): string {
-  if (lane.motion_state === 'stalled') return 'bad'
-  if (lane.hard_flags.some(flag => flag.severity === 'bad')) return 'bad'
-  if (lane.motion_state === 'waiting') return 'warn'
-  if (lane.hard_flags.some(flag => flag.severity === 'warn')) return 'warn'
-  return 'ok'
-}
-
-export function SwarmHealthBar({ lanes }: { lanes: CommandPlaneSwarmLane[] }) {
-  const counts = { moving: 0, waiting: 0, stalled: 0, terminal: 0 }
-  for (const lane of lanes) {
-    const m = lane.motion_state as keyof typeof counts
-    if (m in counts) counts[m]++
-    else counts.waiting++
-  }
-  const total = lanes.length
-  if (total === 0) return null
-
-  const segments: Array<{ key: string; count: number; color: string }> = [
-    { key: 'moving', count: counts.moving, color: 'var(--ok)' },
-    { key: 'waiting', count: counts.waiting, color: 'var(--warn)' },
-    { key: 'stalled', count: counts.stalled, color: 'var(--bad)' },
-    { key: 'terminal', count: counts.terminal, color: '#556' },
-  ]
-
-  return html`
-    <div>
-      <div class="swarm-health-bar">
-        ${segments.filter(s => s.count > 0).map(s => html`
-          <div class="swarm-health-seg ${s.key}" style="flex: ${s.count}"></div>
-        `)}
-      </div>
-      <div class="swarm-health-labels">
-        ${segments.filter(s => s.count > 0).map(s => html`
-          <span class="swarm-health-label">
-            <span class="swarm-health-swatch" style="background: ${s.color}"></span>
-            ${s.count} ${s.key}
-          </span>
-        `)}
-      </div>
-    </div>
-  `
-}
-
-function SwarmWorkerGrid({ total }: { total: number }) {
-  const maxDots = 20
-  const present = Math.min(total, maxDots)
-  const overflow = total > maxDots ? total - maxDots : 0
-  const dots = Array.from({ length: present })
-
-  return html`
-    <div class="swarm-worker-grid">
-      ${dots.map(() => html`<span class="swarm-worker-dot present"></span>`)}
-      ${overflow > 0 ? html`<span class="swarm-worker-count">+${overflow}</span>` : null}
-      <span class="swarm-worker-count">(워커 ${total})</span>
-    </div>
-  `
-}
-
-function SwarmLaneStrip({ lane }: { lane: CommandPlaneSwarmLane }) {
-  const counts = lane.counts ?? {}
-  const tone = swarmLaneTone(lane)
-  const totalWorkers = counts.workers ?? 0
-  const ops = counts.operations ?? 0
-  const dets = counts.detachments ?? 0
-  const totalOps = ops + dets
-  const progressPercent =
-    lane.motion_state === 'moving'
-      ? 84
-      : lane.motion_state === 'waiting'
-        ? 58
-        : lane.motion_state === 'terminal'
-          ? 100
-          : 26
-
-  return html`
-    <article class="swarm-lane-strip ${toneClass(tone)}">
-      <div class="swarm-lane-head">
-        <div class="swarm-lane-head-left">
-          <span class="swarm-motion-dot ${lane.motion_state}"></span>
-          <div>
-            <span class="swarm-lane-kicker">${lane.kind} · ${lane.source_of_truth}</span>
-            <strong>${lane.label}</strong>
-          </div>
-        </div>
-        <div class="command-tag-row">
-          <span class="command-chip ${toneClass(tone)}">${lane.phase}</span>
-          <span class="command-chip ${toneClass(tone)}">${lane.motion_state}</span>
-          <span class="command-chip">${relativeTime(lane.last_movement_at)}</span>
-        </div>
-      </div>
-      <p class="swarm-lane-reason">${lane.movement_reason}</p>
-      <div class="swarm-lane-track">
-        <span class="${toneClass(tone)}" style=${`width:${progressPercent}%`}></span>
-      </div>
-      <div class="swarm-lane-details">
-        <div class="swarm-lane-row">
-          <span class="swarm-lane-row-label">Step</span>
-          <span>${lane.current_step}</span>
-        </div>
-        ${totalWorkers > 0
-          ? html`
-              <div class="swarm-lane-row">
-                <span class="swarm-lane-row-label">워커</span>
-                <${SwarmWorkerGrid} total=${totalWorkers} />
-              </div>
-            `
-          : null}
-        ${totalOps > 0
-          ? html`
-              <div class="swarm-lane-row">
-                <span class="swarm-lane-row-label">흐름</span>
-                <div class="swarm-mini-bar">
-                  <div class="swarm-mini-bar-fill" style="width: ${totalOps > 0 ? Math.round((ops / totalOps) * 100) : 0}%; background: var(--${tone === 'bad' ? 'bad' : tone === 'warn' ? 'warn' : 'ok'})"></div>
-                </div>
-                <span class="swarm-worker-count">작전 ${ops} · 실행체 ${dets}</span>
-              </div>
-            `
-          : null}
-      </div>
-      ${lane.blockers.length > 0
-        ? html`<div class="swarm-lane-blockers">막힘: ${lane.blockers.join(' · ')}</div>`
-        : null}
-      ${lane.hard_flags.length > 0
-        ? html`
-            <div class="swarm-lane-flags">
-              ${lane.hard_flags.map((flag: CommandPlaneSwarmFlag) => html`<span class="command-chip ${toneClass(flag.severity)}">${flag.code}</span>`)}
-            </div>
-          `
-        : null}
-    </article>
-  `
-}
-
-export function SwarmStoryboard({ lanes }: { lanes: CommandPlaneSwarmLane[] }) {
-  const featured = lanes.slice(0, 4)
-  if (featured.length === 0) return null
-  return html`
-    <div class="swarm-storyboard">
-      ${featured.map(lane => {
-        const tone = swarmLaneTone(lane)
-        const workers = lane.counts.workers ?? 0
-        const operations = lane.counts.operations ?? 0
-        const detachments = lane.counts.detachments ?? 0
-        return html`
-          <article class="swarm-story-card ${toneClass(tone)}">
-            <div class="swarm-story-topline">
-              <span class="command-chip ${toneClass(tone)}">${lane.motion_state}</span>
-              <span class="command-chip">${lane.phase}</span>
-            </div>
-            <strong>${lane.label}</strong>
-            <p>${lane.current_step}</p>
-            <div class="swarm-story-strip">
-              <span>워커 ${workers}</span>
-              <span>작전 ${operations}</span>
-              <span>실행체 ${detachments}</span>
-            </div>
-            <small>${lane.movement_reason}</small>
-          </article>
-        `
-      })}
-    </div>
-  `
-}
-
-function SwarmEventNode({ event }: { event: CommandPlaneSwarmTimelineEvent }) {
-  const ts = event.timestamp ? new Date(event.timestamp) : null
-  const validTs = ts && !isNaN(ts.getTime()) ? ts : null
-  const timeStr = validTs ? `${String(validTs.getHours()).padStart(2, '0')}:${String(validTs.getMinutes()).padStart(2, '0')}` : ''
-  return html`
-    <div class="swarm-event-node">
-      <span class="swarm-event-dot ${toneClass(event.tone)}"></span>
-      <span class="swarm-event-time">${timeStr}</span>
-      <div class="swarm-event-body">
-        <strong>${event.title}</strong>
-        <span class="swarm-event-kind">${event.kind}</span>
-        ${event.detail ? html`<div class="command-card-sub">${event.detail}</div>` : null}
-      </div>
-    </div>
-  `
-}
-
-function SwarmGapDot({ gap }: { gap: CommandPlaneSwarmGap }) {
-  return html`
-    <div class="swarm-gap-inline">
-      <span class="swarm-gap-dot"></span>
-      <span class="command-chip ${toneClass(gap.severity)}">${gap.code} (${gap.count})</span>
-      <span class="command-card-sub">${gap.summary}</span>
-    </div>
-  `
-}
-
-function SwarmProofPanel({ proof }: { proof?: CommandPlaneSwarmProof }) {
-  const tone =
-    proof?.status === 'missing'
-      ? 'warn'
-      : proof?.pass === false
-        ? 'bad'
-        : proof?.pass === true
-          ? 'ok'
-          : 'warn'
-  return html`
-    <div class="command-guide-card ${toneClass(tone)}">
-        <div class="command-guide-head">
-          <strong>Hot Proof / 가동 증거</strong>
-          <span class="command-chip ${toneClass(tone)}">${proof?.status ?? 'missing'}</span>
-        </div>
-      ${proof
-        ? html`
-            <div class="command-card-grid">
-              <span>소스</span><span>${proof.source}</span>
-              <span>런</span><span>${proof.run_id ?? 'n/a'}</span>
-              <span>수집 시각</span><span>${relativeTime(proof.captured_at)}</span>
-              <span>통과</span><span>${proof.pass == null ? 'n/a' : proof.pass ? '예' : '아니오'}</span>
-              <span>최대 Hot Slots</span><span>${proof.peak_hot_slots ?? 'n/a'}</span>
-              <span>Ctx / Slot</span><span>${proof.ctx_per_slot ?? 'n/a'}</span>
-              <span>워커 증거</span><span>${proof.workers.expected ?? 'n/a'} 예상 · ${proof.workers.done ?? 'n/a'} 완료 · ${proof.workers.final ?? 'n/a'} 최종</span>
-            </div>
-            ${proof.artifact_ref
-              ? html`<div class="command-card-foot">${proof.artifact_ref}</div>`
-              : null}
-            ${proof.missing_reason
-              ? html`<p>${proof.missing_reason}</p>`
-              : null}
-          `
-        : html`<p>아직 스웜 증거가 수집되지 않았습니다.</p>`}
-    </div>
-  `
-}
+// Re-export for consumers that import from './swarm'
+export { SwarmBlockerCard, SwarmChecklistCard, SwarmWorkerCard } from './swarm-cards'
+export { SwarmHealthBar, SwarmRunResolutionCard, SwarmStoryboard } from './swarm-storyboard'
 
 function SwarmPanel() {
   const summary = currentCommandPlaneSummary()
@@ -476,72 +104,6 @@ function SwarmPanel() {
           `
         : html`<div class="empty-state">스웜 상태를 아직 불러오지 못했습니다.</div>`}
     </section>
-  `
-}
-
-export function SwarmChecklistCard({ item }: { item: CommandPlaneSwarmChecklistItem }) {
-  return html`
-    <article class="command-guide-card ${toneClass(item.status)}">
-      <div class="command-guide-head">
-        <strong>${item.title}</strong>
-        <span class="command-chip ${toneClass(item.status)}">${item.status}</span>
-      </div>
-      <p>${item.detail}</p>
-      <div class="command-card-foot">Next tool: ${item.next_tool}</div>
-    </article>
-  `
-}
-
-export function SwarmBlockerCard({ blocker }: { blocker: CommandPlaneSwarmBlocker }) {
-  return html`
-    <article class="command-alert ${toneClass(blocker.severity)}">
-      <div class="command-card-head">
-        <strong>${blocker.title}</strong>
-        <span class="command-chip ${toneClass(blocker.severity)}">${blocker.severity}</span>
-      </div>
-      <div class="command-alert-meta">
-        <span>${blocker.code}</span>
-        <span>next ${blocker.next_tool}</span>
-      </div>
-      <p>${blocker.detail}</p>
-    </article>
-  `
-}
-
-export function SwarmWorkerCard({ worker }: { worker: CommandPlaneSwarmWorker }) {
-  return html`
-    <article class="command-card compact">
-      <div class="command-card-head">
-        <div>
-          <strong>${worker.name}</strong>
-          <div class="command-card-sub">${worker.role} · ${worker.lane}</div>
-        </div>
-        <span class="command-chip ${toneClass(worker.joined ? (worker.heartbeat_fresh ? 'ok' : 'warn') : 'bad')}">
-          ${worker.status}
-        </span>
-      </div>
-      <div class="command-card-grid">
-        <span>Joined</span><span>${worker.joined ? 'yes' : 'no'}</span>
-        <span>Live</span><span>${worker.live_presence ? 'yes' : 'no'}</span>
-        <span>Completed</span><span>${worker.completed ? 'yes' : 'no'}</span>
-        <span>Task</span><span>${worker.current_task ?? worker.bound_task_id ?? 'none'}</span>
-        <span>Task Title</span><span>${worker.bound_task_title ?? 'n/a'}</span>
-        <span>Task Status</span><span>${worker.bound_task_status ?? 'n/a'}</span>
-        <span>Heartbeat</span><span>${worker.heartbeat_age_sec != null ? `${Math.round(worker.heartbeat_age_sec)}s` : worker.heartbeat_fresh ? 'completed-cleanly' : 'n/a'}</span>
-        <span>Squad</span><span>${worker.squad_member ? 'yes' : 'no'}</span>
-        <span>Detachment</span><span>${worker.detachment_member ? 'yes' : 'no'}</span>
-      </div>
-      <div class="command-tag-row">
-        <span class="command-tag">${worker.lane}</span>
-        <span class="command-tag ${worker.current_task_matches_run ? 'ok' : 'warn'}">current_task</span>
-        <span class="command-tag ${worker.claim_marker_seen ? 'ok' : 'warn'}">claim</span>
-        <span class="command-tag ${worker.done_marker_seen ? 'ok' : 'warn'}">done</span>
-        <span class="command-tag ${worker.final_marker_seen ? 'ok' : 'warn'}">final</span>
-      </div>
-      ${worker.last_message
-        ? html`<div class="command-card-foot">${relativeTime(worker.last_message.timestamp)} · ${worker.last_message.content}</div>`
-        : null}
-    </article>
   `
 }
