+++ dashboard/src/components/mission-cards.ts
@@ -1,10 +1,7 @@
 import { html } from 'htm/preact'
 import { Card } from './common/card'
-import { extractAgentInfo } from './common/agent-info'
-import { linkedRecentToolsEmptyState, observedToolsEmptyState, toolAuditStateLabel } from './common/tool-audit'
 import { ProvenanceStrip } from './common/provenance-strip'
 import { openAgentDetail } from './agent-detail'
-import { openKeeperDetail } from './keeper-detail'
 import {
   missionBriefing,
   missionBriefingError,
@@ -14,22 +11,14 @@ import {
 import { workflowActionLabel } from '../workflow-context'
 import type {
   DashboardMissionAttentionQueueItem,
-  DashboardMissionInternalSignal,
   DashboardMissionSessionBrief,
-  DashboardMissionSessionCard,
-  DashboardMissionSessionDetailResponse,
-  Keeper,
 } from '../types'
 import {
-  type EnrichedAgentRow,
-  type EnrichedKeeperRow,
   toneClass,
   relativeTime,
-  formatDuration,
   statusLabel,
   missionTargetTypeLabel,
   signalClassLabel,
-  trimText,
   actionModeLabel,
   actionTargetLabel,
   toggleAttention,
@@ -38,10 +27,12 @@ import {
   openIncidentCommand,
   openActionIntervene,
   openActionCommand,
-  openSession,
   attentionAsIncident,
-  liveStateClass,
 } from './mission-utils'
+
+// Re-export from split files for consumers importing from './mission-cards'
+export { SessionBriefCard, SessionDetailCard } from './mission-session-cards'
+export { AgentBriefCard, KeeperBriefCard, InternalSignalCard } from './mission-agent-cards'
 
 export function MissionContextBar({
   cluster,
@@ -313,399 +304,6 @@ export function AttentionCard({
               <button class="control-btn ghost" onClick=${() => openIncidentIntervene(incident)}>이 이슈로 개입 열기</button>
               <button class="control-btn ghost" onClick=${() => openIncidentCommand(incident)}>이 이슈의 원인 보기</button>
             `}
-      </div>
-    </article>
-  `
-}
-
-export function SessionBriefCard({
-  brief,
-  selected,
-}: {
-  brief: DashboardMissionSessionCard
-  selected: boolean
-}) {
-  const members = brief.member_previews.slice(0, 4)
-  const action = brief.top_recommendation ?? null
-  const incident = brief.top_attention ?? null
-  const liveCount = brief.active_count ?? 0
-  const seenCount = brief.seen_count ?? liveCount
-  const plannedCount = brief.planned_count ?? brief.member_names.length
-
-  return html`
-    <article class="mission-crew-card ${toneClass(brief.top_attention?.severity ?? brief.health ?? brief.status)} ${liveStateClass(brief.status, brief.health)} ${selected ? 'is-selected' : ''}">
-      <button class="mission-card-select" onClick=${() => toggleSession(brief.session_id)}>
-        <div class="mission-card-head">
-          <div>
-            <div style="display:flex;align-items:center;gap:8px">
-              <div class="mission-status-dot ${liveStateClass(brief.status, brief.health)}"></div>
-              <strong>${brief.goal}</strong>
-            </div>
-            <div class="mission-card-target">${brief.session_id}${brief.room ? ` · ${brief.room}` : ''}</div>
-          </div>
-          <span class="command-chip ${toneClass(brief.top_attention?.severity ?? brief.health ?? brief.status)}">${statusLabel(brief.status)}</span>
-        </div>
-
-        <div class="mission-fact-grid">
-          <div class="mission-fact-tile">
-            <span>멤버</span>
-            <strong>${brief.member_names.length}</strong>
-            <small>${brief.member_names.slice(0, 3).join(', ') || '없음'}</small>
-          </div>
-          <div class="mission-fact-tile">
-            <span>가동 시간</span>
-            <strong>${formatDuration(brief.elapsed_sec)}</strong>
-            <small>${brief.started_at ? `${relativeTime(brief.started_at)} 시작` : '시작 시각 없음'}</small>
-          </div>
-          <div class="mission-fact-tile">
-            <span>최근 흐름</span>
-            <strong>${brief.last_event_at ? relativeTime(brief.last_event_at) : '기록 없음'}</strong>
-            <small>${brief.communication_summary ?? '요약 없음'}</small>
-          </div>
-          <div class="mission-fact-tile">
-            <span>충원 상태</span>
-            <strong>${liveCount}/${brief.required_count || 1}</strong>
-            <small>live · seen ${seenCount} · planned ${plannedCount}</small>
-          </div>
-        </div>
-      </button>
-
-      ${brief.blocker_summary ? html`<div class="mission-inline-note">막힘 · ${brief.blocker_summary}</div>` : null}
-      ${brief.counts_basis ? html`<div class="mission-inline-note">관측 기준 · ${brief.counts_basis}</div>` : null}
-
-      <div class="mission-crew-event">
-        <span>최근 사건</span>
-        <strong>${brief.last_event_summary ?? '최근 세션 이벤트가 없습니다.'}</strong>
-        <small>${brief.last_event_at ? relativeTime(brief.last_event_at) : '시각 없음'}</small>
-      </div>
-
-      ${brief.operation_badges.length > 0
-        ? html`
-            <div class="mission-pill-row">
-              ${brief.operation_badges.slice(0, 3).map(operation => html`
-                <span class="mission-pill">
-                  ${operation.operation_id} · ${statusLabel(operation.status)}${operation.stage ? ` · ${operation.stage}` : ''}
-                </span>
-              `)}
-            </div>
-          `
-        : null}
-
-      ${members.length > 0
-        ? html`
-            <div class="mission-member-preview-grid">
-              ${members.map(member => html`
-                <button class="mission-member-preview" onClick=${() => openAgentDetail(member.agent_name)}>
-                  <strong>${member.agent_name}</strong>
-                  <span>
-                    ${member.current_work ?? '현재 작업 없음'}
-                    ${member.is_live === false ? ' · archived' : member.is_live === true ? ' · live' : ''}
-                  </span>
-                  <small>${member.recent_output_preview ?? member.recent_input_preview ?? '최근 입출력 없음'}</small>
-                </button>
-              `)}
-            </div>
-          `
-        : null}
-
-      <div class="mission-card-actions">
-        <button class="control-btn ghost" onClick=${() => openSession('control', brief.session_id)}>세션 개입 열기</button>
-        <button class="control-btn ghost" onClick=${() => openSession('lab', brief.session_id)}>세션 원인 보기</button>
-        ${action
-          ? html`<button class="control-btn ghost" onClick=${() => openActionIntervene(action, incident, '상황판 세션 요약')}>추천 액션 열기</button>`
-          : null}
-      </div>
-    </article>
-  `
-}
-
-export function SessionDetailCard({
-  detail,
-  loading,
-  error,
-}: {
-  detail: DashboardMissionSessionDetailResponse | null
-  loading: boolean
-  error: string | null
-}) {
-  if (loading && !detail) {
-    return html`
-      <${Card} title="세션 상세" class="mission-list-card">
-        <div class="loading-indicator">세션 상세 불러오는 중...</div>
-      <//>
-    `
-  }
-
-  if (error && !detail) {
-    return html`
-      <${Card} title="세션 상세" class="mission-list-card">
-        <div class="empty-state error">${error}</div>
-      <//>
-    `
-  }
-
-  if (!detail?.session) {
-    return null
-  }
-
-  const session = detail.session
-  return html`
-    <${Card} title="세션 상세" class="mission-list-card" semanticId="mission.session_detail">
-      <div class="mission-section-head">
-        <h3>${session.goal}</h3>
-        <p>${session.session_id}${session.room ? ` · ${session.room}` : ''}</p>
-      </div>
-
-      ${error ? html`<div class="mission-inline-note">${error}</div>` : null}
-
-      <div class="mission-detail-grid">
-        <div class="mission-detail-column">
-          <div class="mission-card-head">
-            <strong>타임라인</strong>
-            <span class="command-chip">${detail.timeline.length}</span>
-          </div>
-          <div class="mission-timeline-list">
-            ${detail.timeline.length > 0
-              ? detail.timeline.map(item => html`
-                  <article class="mission-timeline-row">
-                    <div class="mission-card-head">
-                      <strong>${item.summary}</strong>
-                      <span>${item.timestamp ? relativeTime(item.timestamp) : '시각 없음'}</span>
-                    </div>
-                    <small>${item.actor ? `${item.actor} · ` : ''}${item.event_type ?? '이벤트'}</small>
-                  </article>
-                `)
-              : html`<div class="empty-state">표시할 세션 이벤트가 없습니다.</div>`}
-          </div>
-        </div>
-
-        <div class="mission-detail-column">
-          <div class="mission-card-head">
-            <strong>참여자</strong>
-            <span class="command-chip">${detail.participants.length}</span>
-          </div>
-          <div class="mission-activity-list compact">
-            ${detail.participants.length > 0
-              ? detail.participants.map(participant => html`
-                  <button class="mission-member-preview" onClick=${() => openAgentDetail(participant.agent_name)}>
-                    <strong>${participant.agent_name}</strong>
-                    <span>${participant.current_work ?? '현재 작업 없음'}</span>
-                    <small>
-                      ${participant.recent_output_preview ?? participant.recent_input_preview ?? '최근 입출력 없음'}
-                      ${participant.last_activity_at ? ` · ${relativeTime(participant.last_activity_at)}` : ''}
-                    </small>
-                  </button>
-                `)
-              : html`<div class="empty-state">세션 참여자 미리보기가 없습니다.</div>`}
-          </div>
-        </div>
-      </div>
-
-      <div class="mission-detail-grid">
-        <div class="mission-detail-column">
-          <div class="mission-card-head">
-            <strong>연결된 작전</strong>
-            <span class="command-chip">${detail.operations.length}</span>
-          </div>
-          <div class="mission-link-list">
-            ${detail.operations.length > 0
-              ? detail.operations.map(operation => html`
-                  <button class="mission-link-row" onClick=${() => openSession('lab', session.session_id)}>
-                    <strong>${operation.operation_id}</strong>
-                    <span>${statusLabel(operation.status)}${operation.stage ? ` · ${operation.stage}` : ''}</span>
-                    <small>${operation.detachment_status ?? operation.objective ?? '분견대 정보 없음'}</small>
-                  </button>
-                `)
-              : html`<div class="empty-state">연결된 작전이 없습니다.</div>`}
-          </div>
-        </div>
-
-        <div class="mission-detail-column">
-          <div class="mission-card-head">
-            <strong>연속성 관찰</strong>
-            <span class="command-chip">${detail.keepers.length}</span>
-          </div>
-          <div class="mission-link-list">
-            ${detail.keepers.length > 0
-              ? detail.keepers.map(keeper => html`
-                  <div class="mission-link-row static">
-                    <strong>${keeper.name}</strong>
-                    <span>${statusLabel(keeper.status)}${keeper.generation != null ? ` · 세대 ${keeper.generation}` : ''}</span>
-                    <small>${keeper.current_work ?? '현재 작업 정보 없음'}</small>
-                  </div>
-                `)
-              : html`<div class="empty-state">직접 연결된 키퍼는 없습니다.</div>`}
-          </div>
-        </div>
-      </div>
-    <//>
-  `
-}
-
-export function AgentBriefCard({ row }: { row: EnrichedAgentRow }) {
-  const info = extractAgentInfo(row.brief.agent_name)
-  const who = row.withWhom.length > 0 ? row.withWhom.slice(0, 3).join(', ') : '단독 또는 방 단위'
-  const recentToolsLabel =
-    row.recentTools.length > 0
-      ? row.recentTools.join(', ')
-      : toolAuditStateLabel(observedToolsEmptyState(row.keeper, row.brief.tool_audit_source))
-  return html`
-    <article class="mission-activity-card ${toneClass(row.brief.status ?? row.agent?.status)}">
-      <button class="mission-card-select" onClick=${() => openAgentDetail(row.brief.agent_name)}>
-        <div class="mission-activity-head">
-          <div class="mission-activity-title">
-            <span class="agent-emoji">${row.agent?.emoji ?? row.keeper?.emoji ?? ''}</span>
-            <div>
-              <strong>${row.brief.agent_name}</strong>
-              <span>${info.model !== info.nickname ? `${info.model} · ` : ''}${info.nickname}</span>
-            </div>
-          </div>
-          <span class="command-chip ${toneClass(row.brief.status ?? row.agent?.status)}">${statusLabel(row.brief.status ?? row.agent?.status)}</span>
-        </div>
-
-        <div class="mission-activity-meta">
-          <span>어디서 · ${row.where}</span>
-          <span>누구와 · ${who}</span>
-          <span>주의 신호 · ${row.brief.related_attention_count}</span>
-        </div>
-
-        <div class="mission-activity-focus">
-          <span>무엇을</span>
-          <strong>${row.currentWork}</strong>
-          ${row.how ? html`<small>어떻게 · ${row.how}</small>` : null}
-        </div>
-      </button>
-
-      <details class="mission-card-disclosure">
-        <summary>최근 흐름</summary>
-        <div class="mission-activity-foot">
-          ${row.recentEvent ? html`<span>최근 일 · ${row.recentEvent}</span>` : html`<span>최근 사건 요약 없음</span>`}
-          <span>관련 세션 · ${row.brief.related_session_id ?? '없음'}</span>
-        </div>
-
-        <details class="mission-card-disclosure compact">
-          <summary>입력 · 응답 · 도구</summary>
-          <div class="mission-io-stack">
-            <div class="mission-io-item">
-              <span>최근 입력</span>
-              <strong>${row.recentInput ?? '표시 가능한 최근 입력이 없습니다'}</strong>
-            </div>
-            <div class="mission-io-item">
-              <span>최근 응답</span>
-              <strong>${row.recentOutput ?? '표시 가능한 최근 응답이 없습니다'}</strong>
-            </div>
-          </div>
-          <div class="mission-activity-foot">
-            <span>최근 도구 · ${recentToolsLabel}</span>
-          </div>
-        </details>
-      </details>
-    </article>
-  `
-}
-
-export function KeeperBriefCard({ row }: { row: EnrichedKeeperRow }) {
-  const continuity = [
-    `세대 ${row.brief.generation ?? row.keeper?.generation ?? 0}`,
-    row.brief.context_ratio != null
-      ? `컨텍스트 ${Math.round(row.brief.context_ratio * 100)}%`
-      : (row.keeper?.context_ratio != null ? `컨텍스트 ${Math.round(row.keeper.context_ratio * 100)}%` : null),
-    row.brief.last_turn_ago_s != null ? `최근 턴 ${Math.round(row.brief.last_turn_ago_s)}초 전` : null,
-  ]
-    .filter((value): value is string => value !== null)
-    .join(' · ')
-  const recentToolsLabel =
-    row.recentTools.length > 0
-      ? row.recentTools.join(', ')
-      : toolAuditStateLabel(linkedRecentToolsEmptyState(row.keeper))
-
-  return html`
-    <article class="mission-activity-card ${toneClass(row.brief.status ?? row.keeper?.status)} ${liveStateClass(row.brief.status, row.keeper?.status)}">
-      <button class="mission-card-select" onClick=${() => {
-        const keeper: Keeper = row.keeper ?? {
-          name: row.brief.name,
-          agent_name: row.brief.agent_name ?? row.brief.name,
-          status: row.brief.status ?? 'unknown',
-          context_ratio: row.brief.context_ratio ?? null,
-        } as Keeper
-        openKeeperDetail(keeper)
-      }}>
-        <div class="mission-activity-head">
-          <div class="mission-activity-title">
-            <div class="mission-status-dot ${liveStateClass(row.brief.status, row.keeper?.status)}"></div>
-            <span class="agent-emoji">${row.keeper?.emoji ?? ''}</span>
-            <div>
-              <strong>${row.brief.name}</strong>
-              ${row.keeper?.koreanName ? html`<span>${row.keeper.koreanName}</span>` : null}
-            </div>
-          </div>
-          <span class="command-chip ${toneClass(row.brief.status ?? row.keeper?.status)}">${statusLabel(row.brief.status ?? row.keeper?.status)}</span>
-        </div>
-
-        <div class="mission-activity-meta">
-          <span>최근 하트비트 · ${row.keeper?.last_heartbeat ? relativeTime(row.keeper.last_heartbeat) : '기록 없음'}</span>
-          <span>${continuity || '연속성 정보 없음'}</span>
-        </div>
-
-        <div class="mission-activity-focus">
-          <span>무엇을</span>
-          <strong>${row.currentWork}</strong>
-          ${row.keeper?.skill_reason ? html`<small>판단 요약 · ${trimText(row.keeper.skill_reason, 120)}</small>` : null}
-        </div>
-      </button>
-
-      <details class="mission-card-disclosure">
-        <summary>연속성 상세</summary>
-        <div class="mission-activity-foot">
-          <span>에이전트 · ${row.brief.agent_name ?? row.keeper?.agent_name ?? '기록 없음'}</span>
-          ${row.recentEvent ? html`<span>최근 일 · ${row.recentEvent}</span>` : null}
-        </div>
-        <details class="mission-card-disclosure compact">
-          <summary>입력 · 응답 · 도구</summary>
-          <div class="mission-io-stack">
-            <div class="mission-io-item">
-              <span>최근 입력</span>
-              <strong>${row.recentInput ?? '표시 가능한 최근 입력이 없습니다'}</strong>
-            </div>
-            <div class="mission-io-item">
-              <span>최근 응답</span>
-              <strong>${row.recentOutput ?? '표시 가능한 최근 응답이 없습니다'}</strong>
-            </div>
-          </div>
-          <div class="mission-activity-foot">
-            <span>최근 도구 · ${recentToolsLabel}</span>
-          </div>
-        </details>
-      </details>
-    </article>
-  `
-}
-
-export function InternalSignalCard({ item }: { item: DashboardMissionInternalSignal }) {
-  const action = item.action ?? null
-  const attention = item.attention ?? null
-  return html`
-    <article class="mission-action-card ${toneClass(item.severity)}">
-      <div class="mission-card-head">
-        <span class="command-chip ${toneClass(item.severity)}">
-          ${item.signal_type === 'action' && action ? workflowActionLabel(action.action_type) : attention?.kind ?? '내부 신호'}
-        </span>
-        <span class="mission-card-target">${missionTargetTypeLabel(item.target_type)}${item.target_id ? ` · ${item.target_id}` : ''}</span>
-      </div>
-      <p>${item.summary}</p>
-      ${action ? html`<div class="mission-action-preview">${action.reason}</div>` : null}
-      <div class="mission-card-actions">
-        ${action
-          ? html`
-              <button class="control-btn ghost" onClick=${() => openActionIntervene(action, attention, '상황판 내부 신호')}>이 액션으로 개입 열기</button>
-              <button class="control-btn ghost" onClick=${() => openActionCommand(action, attention, '상황판 내부 신호')}>이 이슈의 원인 보기</button>
-            `
-          : attention
-            ? html`
-                <button class="control-btn ghost" onClick=${() => openIncidentIntervene(attention)}>이 이슈로 개입 열기</button>
-                <button class="control-btn ghost" onClick=${() => openIncidentCommand(attention)}>이 이슈의 원인 보기</button>
-              `
-            : null}
       </div>
     </article>
   `
